requirements.txt
arabic-reshaper==2.1.1
asgiref==3.2.10
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.1
django-crispy-forms==1.9.2
django-filter==2.4.0
django-import-export==2.5.0
django-mathfilters==1.0.0
django-rest-knox==4.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
et-xmlfile==1.0.1
future==0.18.2
html5lib==1.1
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
MarkupPy==1.14
MarkupSafe==1.1.1
numpy==1.20.2
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.2.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.7
psycopg2==2.8.6
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.1
PyPDF2==1.26.0
PyQRCode==1.2.1
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.4.1
reportlab==3.5.65
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
tablib==3.0.0
traitlets==5.0.4
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
yapf==0.31.0
